requirements.txt
# Core dependencies with version ranges for better compatibility
a2wsgi>=1.10.8,<2.0.0
ansio>=0.0.1
browser-use>=0.5.11,<1.0.0
docker>=7.1.0,<8.0.0
duckduckgo-search>=6.1.12,<7.0.0
fastmcp>=2.3.4,<3.0.0
fasta2a>=0.5.0,<1.0.0
flask[async]>=3.0.3,<4.0.0
flask-basicauth>=0.2.0,<1.0.0
flaredantic>=0.1.4,<1.0.0
GitPython>=3.1.43,<4.0.0
inputimeout>=1.0.4,<2.0.0
kokoro>=0.9.2
simpleeval>=1.0.3,<2.0.0
lxml_html_clean>=0.3.1,<1.0.0
markdown>=3.7,<4.0.0
mcp>=1.13.1,<2.0.0
newspaper3k>=0.2.8,<1.0.0
paramiko>=3.5.0,<4.0.0
pypdf>=6.0.0,<7.0.0
python-dotenv>=1.1.0,<2.0.0
pytz>=2024.2,<2025.0
nest-asyncio>=1.6.0,<2.0.0
crontab>=1.0.1,<2.0.0
pathspec>=0.12.1
psutil>=7.0.0
soundfile>=0.13.1,<1.0.0
aiohttp>=3.8.0,<4.0.0
aiofiles>=23.0.0,<24.0.0

# AI/ML dependencies with platform considerations
# FAISS with platform-specific installation
faiss-cpu>=1.11.0,<2.0.0; sys_platform != "darwin" or platform_machine != "arm64"
faiss-cpu>=1.11.0,<2.0.0; sys_platform == "linux"

# LangChain ecosystem with compatible versions
langchain-core>=0.3.49,<1.0.0
langchain-community>=0.3.19,<1.0.0
langchain-unstructured[all-docs]>=0.1.6,<1.0.0

# LLM and embedding dependencies
openai-whisper>=20240930,<20251000.0
sentence-transformers>=3.0.1,<4.0.0
tiktoken>=0.8.0,<1.0.0

# Document processing
unstructured[all-docs]>=0.16.23,<1.0.0
unstructured-client>=0.31.0,<1.0.0

# Web automation
playwright>=1.52.0,<2.0.0

# Additional utilities
webcolors>=24.6.0,<25.0.0

# Type stubs for better development experience
types-requests>=2.31.0
types-PyYAML>=6.0.0
